Adding back temporary variable
Ah, so I removed this but after I now have removed it I remembered why
it existed.

So when we did `database, err := sql.Open(...)` it would actually create
a new local variable named database (which was not the global variable).

That was my reasoning for doing it like this before, so I am just
reverting it back to the way that it was.